requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
asyncer==0.0.8
attrs==25.3.0
auth0-python==4.8.1
autogen==0.8.0
azure-common==1.1.28
azure-core==1.32.0
azure-search-documents==11.5.2
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.14.0
coloredlogs==15.0.1
comm==0.2.2
crewai==0.108.0
crewai-tools==0.38.1
cryptography==44.0.2
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
deprecated==1.2.18
deprecation==2.1.0
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docstring-parser==0.16
durationpy==0.9
embedchain==0.1.128
et-xmlfile==2.0.0
executing==2.2.0
fast-depends==2.4.12
fastapi==0.115.12
fastavro==1.10.0
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
gitdb==4.0.12
gitpython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.2
gptcache==0.1.44
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.8
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
instructor==1.7.7
ipykernel==6.29.5
ipython==8.34.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.8.2
json-repair==0.40.0
json5==0.10.0
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
kubernetes==32.0.1
lancedb==0.21.1
langchain==0.3.21
langchain-cohere==0.3.5
langchain-community==0.3.20
langchain-core==0.3.48
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.7
langsmith==0.3.18
litellm==1.60.2
load-dotenv==0.1.0
mako==1.3.9
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.77
mistune==3.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
notebook==7.3.2
notebook-shim==0.2.4
numpy==2.2.4
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.68.2
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdfminer-six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
portalocker==2.10.1
posthog==3.21.0
pre-commit==4.1.0
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyautogen==0.8.0
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
pydeck==0.9.1
pygments==2.19.1
pyjwt==2.10.1
pylance==0.25.0
pypdf==5.4.0
pypdfium2==4.30.1
pypika==0.48.9
pyproject-hooks==1.2.0
pyright==1.1.398
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
pyyaml==6.0.2
pyzmq==26.2.1
qdrant-client==1.13.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
schema==0.7.7
send2trash==1.8.3
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlalchemy==2.0.39
stack-data==0.6.3
starlette==0.46.1
streamlit==1.44.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
terminado==0.18.1
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli-w==1.2.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250306
typing-extensions==4.13.0
typing-inspect==0.9.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
uv==0.6.10
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.2
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
